Fix additional lint errors.
Fix errors screaming in lint C++ pass
Run another lintrunner pass just incase we get conflicts again

+++ onnxruntime/core/providers/migraphx/migraphx_execution_provider_utils.h
@@ -152,7 +152,7 @@ bool canEvalNodeArgument(const GraphViewer& graph,
   return true;
 }
 
-float ConvertSinglePrecisionIEEE754ToFloat(unsigned long input) {
+float ConvertSinglePrecisionIEEE754ToFloat(uint32_t input) {
   int s = (input >> 31) & 0x01;
   int e = ((input & 0x7f800000) >> 23) - 127;
   int p = -1;
